feat(kobe): standalone worktree management page
Adds a full-window `kobe worktrees` page (opened with `x` from the Tasks
pane sidebar, mirroring `kobe settings`'s tab-opener shape): lists every
local saved project's git worktrees, marks kobe-managed vs adopted and
dirty state, shows whether the branch has reached origin and how long ago
the worktree was created, and requires a second, more severe confirmation
before force-deleting a worktree with uncommitted/untracked changes.

New daemon RPCs `worktree.list` / `worktree.remove` back the page.
Several pre-existing files this change had to touch were already over the
repo's 500-line file-size cap (cli/index.ts, keybindings.ts, tmux.ts,
remote-orchestrator.ts, tasks-pane/host.tsx, handlers.ts) — all six are
split by domain/responsibility in this same commit, since the split and
the new capability are interleaved in the same lines.
