+++ docs/KEYBINDINGS.md
@@ -115,7 +115,7 @@ Normal startup opens directly into the task's tmux session. These keys are insta
 | tmux `prefix space` | tmux prefix | Toggle Zen mode — collapses **every** ChatTab in the session to its engine pane (hides the file/Ops and terminal panes, and the Tasks rail unless Settings → General → "Keep Tasks pane in zen mode" is on). Zen is session-wide and persistent: tabs you switch to stay collapsed, and a newly created ChatTab opens collapsed too. A second press restores exactly the panes zen hid across all tabs. Also reachable via the `zen` chip above the file list (enter-only, since entering hides that pane). |
 | tmux `prefix f` | tmux prefix | Open the prompt-only quick-task page (asks for just a prompt; repo / engine / base branch default from the current task). |
 
-Inside the Tasks pane itself, plain-letter task actions are pane-local: `n` new task, `s` Settings, `u` update page when an update is available, `o` open worktree, `t` toggle task sort (default/manual vs recent), `a` archive/unarchive, `d` delete, `r` title, `b` branch, `v` engine, `?` collapse/expand the keys legend (KV-persisted; clicking the `── keys ──` header does the same), and `[` / `]` Working session vs Archives. `ctrl+p` cycles the global project filter (`all` → each saved project), synced across every task session through the daemon's `ui-prefs` channel, without restoring repo grouping: PROJECTS rows stay visible as main-session entries, while the TASKS section narrows to the selected repo and still composes with `/` search. `Right` (`tasks.focusEngine`) re-focuses the current window's engine pane — the role-tagged `@kobe_role=claude` pane via `claudePaneIdStrict`, the inverse of the ctrl+h jump to the rail — and no-ops outside tmux (no `$TMUX_PANE`). These host-level letters (and Right) gate on BOTH an empty dialog stack and the sidebar's `/`-search being inactive — while a query is being typed they fall through as literal text (Right keeps moving the search input's cursor); move-mode is unaffected, since it only owns `j`/`k`/`enter`/`esc`. Archive/delete also kill the task's cached tmux session when present, because the legacy outer monitor no longer owns that cleanup path.
+Inside the Tasks pane itself, plain-letter task actions are pane-local: `n` new task, `s` Settings, `x` the standalone worktree-management page (`worktrees.open.sidebar` — cross-project audit + delete, no global companion chord since it's a sidebar-launched utility like `n`, not an anywhere-reachable surface like Settings; kept out of the trimmed footer legend, reachable via F1 full help; not `w`/`e`, which `keymap-slot-parity.test.ts` uses as a free-key example for `sidebar.nav` override testing), `u` update page when an update is available, `o` open worktree, `t` toggle task sort (default/manual vs recent), `a` archive/unarchive, `d` delete, `r` title, `b` branch, `v` engine, `?` collapse/expand the keys legend (KV-persisted; clicking the `── keys ──` header does the same), and `[` / `]` Working session vs Archives. `ctrl+p` cycles the global project filter (`all` → each saved project), synced across every task session through the daemon's `ui-prefs` channel, without restoring repo grouping: PROJECTS rows stay visible as main-session entries, while the TASKS section narrows to the selected repo and still composes with `/` search. `Right` (`tasks.focusEngine`) re-focuses the current window's engine pane — the role-tagged `@kobe_role=claude` pane via `claudePaneIdStrict`, the inverse of the ctrl+h jump to the rail — and no-ops outside tmux (no `$TMUX_PANE`). These host-level letters (and Right) gate on BOTH an empty dialog stack and the sidebar's `/`-search being inactive — while a query is being typed they fall through as literal text (Right keeps moving the search input's cursor); move-mode is unaffected, since it only owns `j`/`k`/`enter`/`esc`. Archive/delete also kill the task's cached tmux session when present, because the legacy outer monitor no longer owns that cleanup path.
 
 The Tasks/Ops panes are version-tagged with `@kobe_pane_version`. After an upgrade, `ensureSession` respawns stale kobe-owned panes in place while preserving the engine pane and ChatTab windows. Do not use `kobe reset` as the normal update path; reset is the runtime-recovery fallback for wedged tmux/daemon state.
