+++ packages/kobe/src/tui/context/keybindings.ts
@@ -70,6 +70,9 @@
 
 import { createSignal } from "solid-js"
 import type { Binding } from "../lib/keymap"
+import { CHAT_BINDINGS } from "./keybindings-chat.ts"
+import { FILES_BINDINGS } from "./keybindings-files.ts"
+import { SIDEBAR_BINDINGS } from "./keybindings-sidebar.ts"
 
 /** Pane scopes used to gate where a binding is active. */
 export type KobeBindingScope = "global" | "sidebar" | "workspace" | "files" | "terminal"
@@ -174,6 +177,19 @@ export const KobeKeymap: readonly KobeBinding[] = [
     hint: { keys: "s", label: "settings", status: false },
   },
   {
+    // Sidebar-only, like `task.new` — a sidebar-launched utility page, not
+    // an anywhere-reachable surface like Settings, so no `ctrl+…` global
+    // companion chord. NOT `w`/`e` — `keymap-slot-parity.test.ts` documents
+    // those two as a free-key example for `sidebar.nav` override testing;
+    // `x` avoids clobbering that.
+    id: "worktrees.open.sidebar",
+    scope: "sidebar",
+    keys: ["x"],
+    category: "Sidebar",
+    description: "Open worktrees",
+    hint: { keys: "x", label: "worktrees", status: false },
+  },
+  {
     // Sidebar-only — single letter `q`. ctrl+q is reserved for
     // "back to sidebar" (focus.sidebar) so the user has a one-chord
     // path out of the composer; once back on the sidebar, `q` is the
@@ -234,583 +250,19 @@ export const KobeKeymap: readonly KobeBinding[] = [
     category: "Workspace",
     description: "Interrupt current turn (esc while streaming)",
   },
-  // ─── Sidebar ──────────────────────────────────────────────────────────
-  {
-    // POSITIONAL: alternating [down, up] pairs — slot dispatch
-    // (SLOT_CONTRACTS in lib/keymap-overrides.ts). Overrides may supply
-    // any even chord count, e.g. `sidebar.nav: [w, s]`.
-    id: "sidebar.nav",
-    scope: "sidebar",
-    keys: ["j", "k", "down", "up"],
-    category: "Sidebar",
-    description: "Move cursor up/down",
-    hint: { keys: "j/k", label: "nav" },
-  },
-  {
-    id: "sidebar.select",
-    scope: "sidebar",
-    keys: ["return"],
-    category: "Sidebar",
-    description: "Open the selected task",
-    hint: { keys: "enter", label: "select" },
-  },
-  {
-    id: "sidebar.goto",
-    scope: "sidebar",
-    keys: ["g"],
-    category: "Sidebar",
-    description: "Top / bottom of list (gg or shift-G)",
-  },
-  {
-    id: "sidebar.rename",
-    scope: "sidebar",
-    keys: ["r"],
-    category: "Sidebar",
-    description: "Rename task",
-    hint: { keys: "r", label: "rename", status: false },
-  },
-  {
-    id: "sidebar.archive",
-    scope: "sidebar",
-    keys: ["a"],
-    category: "Sidebar",
-    description: "Toggle archive",
-    hint: { keys: "a", label: "archive", status: false },
-  },
-  {
-    id: "sidebar.localMerge",
-    scope: "sidebar",
-    keys: ["m"],
-    category: "Sidebar",
-    description: "Reorder task (Shift+M, then j/k)",
-    hint: { keys: "M", label: "reorder", status: false },
-  },
-  {
-    // Capital P pins / unpins a regular task. Lowercase `p` falls
-    // through to a no-op (the handler gates on evt.shift) so a
-    // mistyped lowercase doesn't churn the flag. Pinned regular tasks
-    // float to the top of the sidebar's flat list, just below the
-    // saved-repo "main" rows. `kind: "main"` rows ignore the chord —
-    // they're implicitly pinned.
-    id: "sidebar.pin",
-    scope: "sidebar",
-    keys: ["p"],
-    category: "Sidebar",
-    description: "Pin / unpin task at top (Shift+P)",
-    hint: { keys: "P", label: "pin", status: false },
-  },
-  {
-    // `i` opens the cursor task in a live read-only preview (the `kobe history`
-    // renderer tailing the transcript) in the engine pane slot instead of the
-    // engine, and toggles back on a second press. For inspecting a task an agent
-    // is working in without driving it. Same beta gate as the archived preview.
-    id: "sidebar.previewToggle",
-    scope: "sidebar",
-    keys: ["i"],
-    category: "Sidebar",
-    description: "Toggle live preview for task (i)",
-    hint: { keys: "i", label: "preview", status: false },
-  },
-  {
-    // POSITIONAL: [previous view, next view] pairs (slot dispatch).
-    id: "sidebar.view",
-    scope: "sidebar",
-    keys: ["[", "]"],
-    category: "Sidebar",
-    description: "Switch view (Working session ↔ Archives)",
-    hint: { keys: "[/]", label: "view", status: false },
-  },
-  {
-    id: "sidebar.sort",
-    scope: "sidebar",
-    keys: ["t"],
-    category: "Sidebar",
-    description: "Switch task sort (default ↔ recent)",
-    hint: { keys: "t", label: "sort", status: false },
-  },
-  {
-    id: "sidebar.projectFilter",
-    scope: "sidebar",
-    keys: ["ctrl+p"],
-    category: "Sidebar",
-    description: "Cycle task project filter",
-    hint: { keys: "ctrl+p", label: "project", status: false },
-  },
-  {
-    id: "sidebar.delete",
-    scope: "sidebar",
-    keys: ["d"],
-    category: "Sidebar",
-    description: "Delete task (with confirm)",
-    hint: { keys: "d", label: "delete", status: false },
-  },
-  {
-    // `/`-search filter. Enters an inline search mode rendered at the
-    // top of the sidebar: typed text fuzz-matches against task title +
-    // repo basename, up/down navigates the filtered list, enter selects
-    // + exits, esc cancels + restores. While search is active the
-    // single-letter sidebar chords (j/k/g/G/d/a/r/P/m) are
-    // de-registered so they fall through to the input as literal text.
-    // `[` / `]` view switch keeps working so the user can search inside
-    // Archives.
-    id: "sidebar.search.enter",
-    scope: "sidebar",
-    keys: ["/"],
-    category: "Sidebar",
-    description: "Search tasks (fuzzy filter)",
-    hint: { keys: "/", label: "search" },
-  },
-  {
-    // Search-mode nav. Only fires while the search input is focused —
-    // j/k are intentionally NOT bound here so they reach the input.
-    // POSITIONAL: [down, up] pairs (slot dispatch).
-    id: "sidebar.search.nav",
-    scope: "sidebar",
-    keys: ["down", "up"],
-    category: "Sidebar",
-    description: "Move highlight in search results",
-  },
-  {
-    // Search-mode submit: select highlighted match and leave search.
-    id: "sidebar.search.submit",
-    scope: "sidebar",
-    keys: ["return"],
-    category: "Sidebar",
-    description: "Select search match and exit search",
-  },
-  {
-    // Search-mode cancel. Only registered while searching; outside
-    // search there is no sidebar-scope esc handler.
-    id: "sidebar.search.cancel",
-    scope: "sidebar",
-    keys: ["escape"],
-    category: "Sidebar",
-    description: "Cancel search (restore prior selection)",
-  },
+  // ─── Sidebar + Tasks pane ─────────────────────────────────────────────
+  // Moved to keybindings-sidebar.ts (file-size cap) — same entries, same
+  // order, same live-binding contract (`kobe tasks` consumes these via
+  // `bindByIds`, following user overrides).
+  ...SIDEBAR_BINDINGS,
 
-  // ─── Tasks pane ───────────────────────────────────────────────────────
-  // The standalone Tasks pane (`kobe tasks`, src/tui/tasks-pane/host.tsx)
-  // consumes these ids via `bindByIds` (since the keybindings-customization
-  // pass; they were raw `{ key: "…" }` literals before), so the rows are
-  // LIVE bindings there and follow user overrides from
-  // `~/.kobe/settings/keybindings.yaml`. New-task (n), settings (s),
-  // rename (r), archive (a), delete (d), merge (M), views ([/]), sort (t)
-  // are already covered by the Sidebar / Global rows above and aren't
-  // duplicated here.
-  {
-    id: "tasks.openWorktree",
-    scope: "sidebar",
-    keys: ["o"],
-    category: "Tasks pane",
-    description: "Open selected task's worktree in your editor",
-    hint: { keys: "o", label: "open wt", status: false },
-  },
-  {
-    id: "tasks.renameBranch",
-    scope: "sidebar",
-    keys: ["b"],
-    category: "Tasks pane",
-    description: "Rename the selected task's git branch",
-    hint: { keys: "b", label: "branch", status: false },
-  },
-  {
-    id: "tasks.cycleEngine",
-    scope: "sidebar",
-    keys: ["v"],
-    category: "Tasks pane",
-    description: "Cycle engine vendor (claude ↔ codex ↔ …) — applies on reopen",
-    hint: { keys: "v", label: "engine", status: false },
-  },
-  {
-    id: "tasks.update",
-    scope: "sidebar",
-    keys: ["u"],
-    category: "Tasks pane",
-    description: "Open the update page (when a new version is available)",
-    hint: { keys: "u", label: "update", status: false },
-  },
-  {
-    // Right arrow jumps from the Tasks pane back into the current
-    // window's engine (claude/codex) pane — the spatial "go right into
-    // the conversation" gesture, the inverse of ctrl+h. Named key, not a
-    // bare letter, but still sidebar-scoped per the boundary rule; the
-    // Tasks-pane host gates it on no dialog + `/`-search inactive, so
-    // Right typed while searching keeps moving the input cursor.
-    id: "tasks.focusEngine",
-    scope: "sidebar",
-    keys: ["right"],
-    category: "Tasks pane",
-    description: "Focus the engine pane of the current window",
-    hint: { keys: "→", label: "engine", status: false },
-  },
-  {
-    // `?` (shift+/ — terminals deliver the literal character) folds the
-    // Tasks pane's `── keys ──` legend down to its header line and back.
-    // The legend is ~20 rows tall with the tmux session chords included;
-    // on short terminals it crowds out the task list. The collapsed state
-    // persists via KV so the preference survives pane respawns.
-    id: "tasks.toggleKeys",
-    scope: "sidebar",
-    keys: ["?"],
-    category: "Tasks pane",
-    description: "Collapse / expand the keys legend",
-    hint: { keys: "?", label: "fold keys", status: false },
-  },
-
-  // ─── Workspace (tmux) ─────────────────────────────────────────────────
-  // tmux-handover chords that drive the task SESSION (windows/tabs/detach),
-  // not opentui bindings. Listed here so the HelpDialog advertises them; the
-  // bracket / ctrl rows below in "Workspace (chat)" register the real opentui
-  // handlers, while the `prefix f` quick-task is a tmux key-table binding the
-  // session installs (not registered through this keymap at all) — DISPLAY
-  // ONLY, with `keys: []` so nothing tries to bind a literal "prefix f" chord.
-  {
-    id: "tmux.quickTask",
-    scope: "global",
-    keys: [],
-    category: "Workspace (tmux)",
-    description: "Quick new task (tmux prefix, then f)",
-    hint: { keys: "prefix f", label: "new task", status: false },
-  },
-  {
-    id: "tmux.engineTab",
-    scope: "global",
-    keys: [],
-    category: "Workspace (tmux)",
-    description: "Open the engine in a new tab (tmux prefix, then t)",
-    hint: { keys: "prefix t", label: "engine tab", status: false },
-  },
-  {
-    id: "tmux.layout.workspaceSplit",
-    scope: "global",
-    keys: [],
-    category: "Workspace (tmux)",
-    description: "Add a temporary workspace split (tmux prefix, then s)",
-    hint: { keys: "prefix s", label: "split", status: false },
-  },
-  {
-    id: "tmux.layout.workspaceClose",
-    scope: "global",
-    keys: [],
-    category: "Workspace (tmux)",
-    description: "Close the focused temporary workspace split (tmux prefix, then x)",
-    hint: { keys: "prefix x", label: "close split", status: false },
-  },
-  {
-    id: "tmux.layout.workspaceReset",
-    scope: "global",
-    keys: [],
-    category: "Workspace (tmux)",
-    description: "Close all temporary workspace splits (tmux prefix, then r)",
-    hint: { keys: "prefix r", label: "reset splits", status: false },
-  },
-  {
-    id: "tmux.layout.tasksToggle",
-    scope: "global",
-    keys: [],
-    category: "Workspace (tmux)",
-    description: "Hide / restore the Tasks pane (tmux prefix, then a)",
-    hint: { keys: "prefix a", label: "tasks pane", status: false },
-  },
-  {
-    id: "tmux.layout.opsToggle",
-    scope: "global",
-    keys: [],
-    category: "Workspace (tmux)",
-    description: "Toggle the file/Ops pane (tmux prefix, then o)",
-    hint: { keys: "prefix o", label: "file pane", status: false },
-  },
-  {
-    id: "tmux.layout.terminalToggle",
-    scope: "global",
-    keys: [],
-    category: "Workspace (tmux)",
-    description: "Hide / restore the terminal pane (tmux prefix, then z)",
-    hint: { keys: "prefix z", label: "terminal", status: false },
-  },
-  {
-    id: "tmux.layout.zenToggle",
-    scope: "global",
-    keys: [],
-    category: "Workspace (tmux)",
-    description: "Zen mode — collapse to the engine pane (tmux prefix, then space)",
-    hint: { keys: "prefix space", label: "zen", status: false },
-  },
-
-  // ─── Workspace (chat) ─────────────────────────────────────────────────
-  {
-    // Composer textarea handles enter via its own onKeyDown. This row
-    // exists only for help-dialog + status-bar visibility; no chord is
-    // registered here.
-    id: "chat.send",
-    scope: "workspace",
-    keys: [],
-    category: "Workspace",
-    description: "Send message (composer)",
-    hint: { keys: "enter", label: "send" },
-  },
-  {
-    // Composer textarea inserts a literal newline on shift+enter (kitty/
-    // CSI-u terminals) and ctrl+J everywhere else; no chord is registered
-    // here. Surfaced in the status bar so the user doesn't have to memorize
-    // it after we stripped the inline footer hint from the composer.
-    id: "chat.newline",
-    scope: "workspace",
-    keys: [],
-    category: "Workspace",
-    description: "Newline in composer",
-    hint: { keys: "shift+enter", label: "newline" },
-  },
-  {
-    // Shift+tab inside the composer cycles the per-task permission mode
-    // (default ↔ plan); the chord is registered in
-    // Composer's onKeyDown, not here. Doc-only entry so the status bar
-    // advertises the binding to a focused user.
-    id: "chat.cycle-mode",
-    scope: "workspace",
-    keys: [],
-    category: "Workspace",
-    description: "Cycle permission mode (composer)",
-    hint: { keys: "shift+tab", label: "mode" },
-  },
-  {
-    // Ctrl+enter mid-stream interrupts the in-flight subprocess and
-    // dispatches the new buffer immediately. Plain enter while
-    // streaming queues instead. Chord is registered in Composer's
-    // onKeyDown; this entry is doc-only.
-    id: "chat.steer",
-    scope: "workspace",
-    keys: [],
-    category: "Workspace",
-    description: "Steer (interrupt + send) — mid-stream only",
-    hint: { keys: "ctrl+enter", label: "steer", status: false },
-  },
-  {
-    id: "chat.tab.new",
-    scope: "workspace",
-    keys: ["ctrl+t"],
-    category: "Workspace",
-    description: "New chat tab",
-    hint: { keys: "ctrl+t", label: "new tab" },
-  },
-  {
-    // KOB-74. Quick-fork: from a focused chat tab, spin up a child
-    // task that inherits repo + branch + model from the source. The
-    // dialog asks only for a prompt; the fork's first turn fires
-    // immediately. `ctrl+t` is taken by `chat.tab.new` (same task,
-    // new tab) so the requested `ctrl+shift+t` would collide — the
-    // keymap layer drops `shift+` on letter keys (terminals deliver
-    // shift+letter as uppercase, not as a modifier event), making
-    // `ctrl+shift+t` and `ctrl+t` indistinguishable at match time.
-    // Picked `ctrl+f` ("fork") because it's free across the keymap,
-    // ctrl+letter has stable C0 byte mappings that work in every
-    // terminal, and the workspace scope keeps it from intruding on
-    // other panes. See docs/KEYBINDINGS.md decision log.
-    id: "chat.fork.new",
-    scope: "workspace",
-    keys: ["ctrl+f"],
-    category: "Workspace",
-    description: "Quick-fork: create child task seeded with current repo/branch/model",
-    hint: { keys: "ctrl+f", label: "fork" },
-  },
-  {
-    // Mirror of claude-code's `/resume` slash. Pops a picker listing
-    // every persisted session for the active task's worktree so the
-    // user can jump back to (or fork from) any prior conversation.
-    // Selecting an already-open session focuses its tab; otherwise a
-    // new tab is opened seeded with that sessionId. Chord chosen for
-    // mnemonic "yank from history" — `ctrl+r` belongs to the prompt
-    // history palette (claude-code parity, KOB-154) and `ctrl+h`
-    // collides with terminals' backspace byte.
-    id: "chat.session.resume",
-    scope: "workspace",
-    keys: ["ctrl+y"],
-    category: "Workspace",
-    description: "Resume a prior session in this task's worktree",
-    hint: { keys: "ctrl+y", label: "resume", status: false },
-  },
-  {
-    id: "chat.tab.close",
-    scope: "workspace",
-    keys: ["ctrl+w"],
-    category: "Workspace",
-    description: "Close chat tab",
-    hint: { keys: "ctrl+w", label: "close tab", status: false },
-  },
-  {
-    // Rename the active chat tab. F2 is the cross-OS / cross-IDE
-    // rename convention (file managers on Windows + Linux, IntelliJ,
-    // VS Code etc.) — chosen here because `ctrl+r` is owned by the
-    // composer's prompt-history palette (claude-code parity, KOB-154
-    // → KOB-156). F2 has no other binding in kobe and doesn't
-    // collide with terminal bytes the way some control chords do.
-    id: "chat.tab.rename",
-    scope: "workspace",
-    keys: ["f2"],
-    category: "Workspace",
-    description: "Rename active chat tab",
-    hint: { keys: "f2", label: "rename tab", status: false },
-  },
-  {
-    // `ctrl+]` cycles forward, `ctrl+[` cycles backward — bracket
-    // pair mirrors the sidebar's `[/]` view switcher and the files
-    // pane's `[/]` tab cycler so the bracket-pair pattern is
-    // consistent across panes. The earlier `ctrl+tab` /
-    // `ctrl+shift+tab` chord is dropped: `tab` is the global
-    // pane-cycle (focus.next) and the ctrl-prefixed variant felt
-    // collision-prone.
-    id: "chat.tab.cycle-next",
-    scope: "workspace",
-    keys: ["ctrl+]"],
-    category: "Workspace",
-    description: "Next chat tab",
-    hint: { keys: "ctrl+]", label: "next tab", status: false },
-  },
-  {
-    id: "chat.tab.cycle-prev",
-    scope: "workspace",
-    keys: ["ctrl+["],
-    category: "Workspace",
-    description: "Previous chat tab",
-    hint: { keys: "ctrl+[", label: "prev tab", status: false },
-  },
-  // AskUserQuestion picker bindings — only fire when a question card is
-  // up (QuestionRow gates `enabled` on its own state). j/k/space/enter/
-  // 1-9 are bare-letter chords by intent: while a picker is showing, the
-  // composer is hidden (Chat.tsx `<Show when={!pendingQuestion()}>`) so
-  // these never compete with composer typing. Workspace scope means the
-  // chat pane must own focus — the user can still navigate the file tree
-  // with j/k while a question is queued.
-  //
-  // NOTE: with the tmux-native model the legacy Chat pane (and its
-  // QuestionRow) is gone, so these rows currently have NO live
-  // registration site — they're display-only. `chat.question.nav` /
-  // `chat.question.pick-number` stay in FIXED_BINDING_IDS for that
-  // reason: an override would change Help copy without changing
-  // behavior. If the picker returns, implement its nav with slot
-  // dispatch (see sidebar.nav) before unlocking them.
-  {
-    id: "chat.question.nav",
-    scope: "workspace",
-    keys: ["j", "k", "down", "up"],
-    category: "Workspace",
-    description: "Move highlight in question picker",
-    hint: { keys: "j/k", label: "pick", status: false },
-  },
-  {
-    id: "chat.question.toggle",
-    scope: "workspace",
-    keys: ["space"],
-    category: "Workspace",
-    description: "Toggle highlighted option in question picker",
-    hint: { keys: "space", label: "toggle", status: false },
-  },
-  {
-    id: "chat.question.submit",
-    scope: "workspace",
-    keys: ["return"],
-    category: "Workspace",
-    description: "Advance / submit question picker",
-    hint: { keys: "enter", label: "submit", status: false },
-  },
-  {
-    id: "chat.question.pick-number",
-    scope: "workspace",
-    keys: ["1", "2", "3", "4", "5", "6", "7", "8", "9"],
-    category: "Workspace",
-    description: "Pick option by number in question picker",
-    hint: { keys: "1-9", label: "pick", status: false },
-  },
+  // ─── Workspace (tmux) + Workspace (chat) ─────────────────────────────
+  // Moved to keybindings-chat.ts (file-size cap) — same entries, same order.
+  ...CHAT_BINDINGS,
 
   // ─── Files ────────────────────────────────────────────────────────────
-  {
-    // POSITIONAL: alternating [down, up] pairs (slot dispatch).
-    id: "files.nav",
-    scope: "files",
-    keys: ["j", "k", "down", "up"],
-    category: "Files",
-    description: "Move cursor up/down",
-    hint: { keys: "j/k", label: "nav" },
-  },
-  {
-    // `h`/`l` for hierarchy navigation in the All tab tree:
-    //   l → expand directory / descend into first child / open file
-    //   h → collapse directory / jump to parent
-    // Plain letters are pane-scoped per the keybinding-boundaries
-    // rule (docs/KEYBINDINGS.md): files-focused only, so they don't
-    // collide with composer typing.
-    // POSITIONAL: alternating [collapse, expand] pairs (slot dispatch).
-    id: "files.hierarchy",
-    scope: "files",
-    keys: ["h", "l", "left", "right"],
-    category: "Files",
-    description: "Collapse / expand tree level",
-    hint: { keys: "h/l", label: "level" },
-  },
-  {
-    // enter → one-key "just open it": opens the file in the user's
-    // nvim/vim (side-by-side `nvim -d` diff vs HEAD when changed, a plain
-    // editable open otherwise), falling back to our own opentui read-only
-    // preview only when no nvim/vim is installed.
-    id: "files.open",
-    scope: "files",
-    keys: ["return"],
-    category: "Files",
-    description: "Open file in nvim (diff vs HEAD when changed)",
-    hint: { keys: "enter", label: "open" },
-  },
-  {
-    // `[` / `]` cycle the All / Changes tabs. Bracket pair matches
-    // the sidebar's Working/Archives view-switcher so the muscle
-    // memory is consistent across panes.
-    // POSITIONAL: [previous tab, next tab] pairs (slot dispatch).
-    id: "files.tab",
-    scope: "files",
-    keys: ["[", "]"],
-    category: "Files",
-    description: "Switch tab (cycle All / Changes)",
-    hint: { keys: "[/]", label: "tab", status: false },
-  },
-  {
-    id: "files.refresh",
-    scope: "files",
-    keys: ["r"],
-    category: "Files",
-    description: "Refresh",
-    hint: { keys: "r", label: "refresh", status: false },
-  },
-  {
-    id: "files.openExternal",
-    scope: "files",
-    keys: ["o"],
-    category: "Files",
-    description: "Open file in system default app (audio / video / pdf preview)",
-    hint: { keys: "o", label: "open external", status: false },
-  },
-  {
-    // `a` → inject `@<path>` into the engine (claude/codex) pane via
-    // tmux send-keys (KOB-232). Enter stays the full-width preview; this
-    // is the "add as a mention" action. Plain letter, files-scoped per
-    // the keybinding-boundaries rule, so it can't collide elsewhere.
-    id: "files.mention",
-    scope: "files",
-    keys: ["a"],
-    category: "Files",
-    description: "Inject @<path> mention into the engine pane",
-    hint: { keys: "a", label: "@mention" },
-  },
-  {
-    // Ops-pane action on the Changes tab. This is the v0.5 Create PR
-    // button rehomed into the file-changes surface: pressing `p` sends
-    // the PR prompt into the engine pane instead of rendering an outer
-    // monitor button.
-    id: "files.createPR",
-    scope: "files",
-    keys: ["p"],
-    category: "Files",
-    description: "Create PR from the current task",
-    hint: { keys: "p", label: "create PR" },
-  },
+  // Moved to keybindings-files.ts (file-size cap) — same entries, same order.
+  ...FILES_BINDINGS,
 
   // ─── Terminal ─────────────────────────────────────────────────────────
   {
