+++ packages/kobe/src/tui/tasks-pane/host.tsx
@@ -1,89 +1,36 @@
 /**
- * `kobe tasks` — the experimental Tasks pane on the far left of a
- * task's tmux session (KOB-233 experiment).
+ * `kobe tasks` — the experimental Tasks pane on the far left of a task's
+ * tmux session (KOB-233). agent-deck-style: keep the task list visible
+ * inside a tmux Session so you can jump between tasks without detaching to
+ * the outer monitor. Reuses the real `Sidebar`; Enter `switch-client`s to
+ * a task's session.
  *
- * agent-deck-style: keep the task list visible while you're inside a
- * tmux Session, so you can jump between tasks without detaching to the
- * outer monitor. Reuses the real `Sidebar`. Enter on a task
- * `tmux switch-client`s to that task's session.
+ * Scope (deliberately minimal): reads `~/.kobe/tasks.json` + re-reads on a
+ * timer as a fallback (the daemon-backed RemoteOrchestrator subscribe is
+ * primary — see `setup.tsx`). `r`/`b`/`v` rename title/branch/cycle engine
+ * via daemon RPCs. `n` opens the same NewTaskDialog + `task.create` RPC as
+ * the outer app. Enter/click lazily creates the session (`ensureSession`,
+ * materialising the worktree first if needed) then switches to it.
  *
- * Scope of the experiment (deliberately minimal):
- *   - Reads `~/.kobe/tasks.json` directly and re-reads on a timer. The
- *     outer app's Orchestrator / Daemon still own most writes; this pane
- *     wires settings / delete / archive to the same daemon-backed paths
- *     as the outer monitor.
- *   - Rename: `r` renames the title (`task.rename` RPC); `b` renames the
- *     branch (`task.setBranch` RPC — `git branch -m` for a materialised
- *     worktree, else just recorded for the eventual ensureWorktree).
- *   - Engine: `v` cycles the task's vendor (`task.setVendor` RPC). Takes
- *     effect on next enter — ensureSession rebuilds the session when its
- *     `@kobe_vendor` tag no longer matches, launching the new engine.
- *   - Create: `n` opens the SAME
- *     NewTaskDialog as the outer app (repo picker + base branch + clone
- *     tab) and fires the daemon's `task.create` RPC — the first
- *     write-path here, the step toward retiring the outer "page 1".
- *     Default repo = cursor task's repo. Backlog task, worktree lazy.
- *   - Switch + lazy-create: Enter / click `switch-client`s to a task's
- *     session, creating it on demand (`ensureSession`) when the task's
- *     worktree already exists on disk — that covers every `main` task
- *     (worktree = repo root) and any worktree task entered at least
- *     once. A backlog task whose worktree was never materialised needs
- *     `git worktree add` from the Orchestrator (which this standalone
- *     pane doesn't have), so it stays a no-op; enter it from the outer
- *     monitor instead.
+ * File-size-cap split: `tasks-pane/actions.ts` (action bodies, deps-bag
+ * template), `tasks-pane/setup.tsx` (boot wiring), `tasks-pane/shortcut-hints.tsx`
+ * (the footer legend) — this file keeps `TasksShell` + thin wrappers.
  */
 
-import { existsSync } from "node:fs"
-import { stat } from "node:fs/promises"
-import {
-  claudePaneIdStrict,
-  currentSessionName,
-  killSession,
-  runTmux,
-  runTmuxCapturing,
-  tmuxSessionName,
-} from "@/tmux/client"
+import { currentSessionName, runTmuxCapturing } from "@/tmux/client"
 import { ZEN_HIDDEN_PANES_OPTION } from "@/tmux/session-layout"
-import { t } from "@/tui/i18n"
-import { TextAttributes } from "@opentui/core"
 import { useTerminalDimensions } from "@opentui/solid"
-import { logClient, logClientError } from "@sma1lboy/kobe-daemon/client/client-log"
-import { connectIfRunning } from "@sma1lboy/kobe-daemon/client/daemon-process"
-import {
-  type Accessor,
-  For,
-  Show,
-  createEffect,
-  createMemo,
-  createSignal,
-  on,
-  onCleanup,
-  onMount,
-  untrack,
-} from "solid-js"
-import { RemoteOrchestrator } from "../../client/remote-orchestrator.ts"
-import { homeDir } from "../../env.ts"
-import { TaskIndexStore } from "../../orchestrator/index/store.ts"
-import { getCustomEngineIds, getPersistedString, setPersistedString } from "../../state/repos.ts"
-import { TMUX_FOCUS_DEFAULTS, resolveUserTmuxKeys } from "../../tmux/keybindings.ts"
+import { type Accessor, createEffect, createMemo, createSignal, on, onCleanup, untrack } from "solid-js"
+import type { RemoteOrchestrator } from "../../client/remote-orchestrator.ts"
 import type { Task } from "../../types/task.ts"
-import { resolvePersistedVendor } from "../../types/vendor.ts"
 import { CURRENT_VERSION, type UpdateInfo } from "../../version.ts"
-import { HelpDialog } from "../component/help-dialog"
-import { NewTaskDialog } from "../component/new-task-dialog"
-import { RenameTaskDialog } from "../component/rename-task-dialog"
-import { SettingsDialog } from "../component/settings-dialog"
-import { ToastOverlay } from "../component/toast-overlay"
 import { VersionSkewBanner } from "../component/version-skew-banner"
-import { bindByIds, findBinding, keymapVersion } from "../context/keybindings"
+import { bindByIds } from "../context/keybindings"
 import { useKV } from "../context/kv"
 import { useNotifications } from "../context/notifications"
 import { useTheme } from "../context/theme"
-import { sessionAttached } from "../lib/attach-gate"
-import { formatChord, tmuxPrefixGlyph } from "../lib/chord-glyphs"
-import { type HostScreen, bootPaneHost } from "../lib/host-boot"
+import { bootPaneHost } from "../lib/host-boot"
 import { useBindings } from "../lib/keymap"
-import { DEFAULT_SETTINGS_SURFACE, SETTINGS_SURFACE_KEY, normalizeSettingsSurface } from "../lib/settings-surface"
 import {
   type CreateTaskContext,
   archiveTaskFlow,
@@ -93,40 +40,29 @@ import {
   renameBranchFlow,
   renameTaskFlow,
 } from "../lib/task-actions"
-import { HandoverError, enterTask } from "../lib/task-enter.ts"
-import { detectWorktreeOpener, openWorktree } from "../lib/worktree-opener"
-import { Sidebar, approxCellWidth } from "../panes/sidebar/Sidebar"
+import { Sidebar } from "../panes/sidebar/Sidebar"
 import type { TaskSortMode } from "../panes/sidebar/groups"
 import { runLayoutAction } from "../panes/terminal/layout-actions.ts"
-import {
-  openHelpTab,
-  openNewTaskTab,
-  openSettingsTab,
-  openUpdateTab,
-  refreshKobeWorkspacePanes,
-} from "../panes/terminal/tmux.ts"
 import { useDialog } from "../ui/dialog"
-import { DialogConfirm } from "../ui/dialog-confirm"
+import {
+  type SwitchToRef,
+  type TasksHostActionsContext,
+  buildTaskActionsContext,
+  focusEnginePaneAction,
+  moveTaskAction,
+  openHelpAction,
+  openSelectedWorktreeAction,
+  openSettingsAction,
+  openUpdateAction,
+  openWorktreesAction,
+  switchToAction,
+  togglePinAction,
+  togglePreviewFlowAction,
+} from "./actions.ts"
+import { setupTasksPane } from "./setup.tsx"
+import { ShortcutHints } from "./shortcut-hints.tsx"
 
-const RELOAD_MS = 1500
-
-/**
- * Toast text for a failed `ensureWorktree`. The new-task dialog blocks
- * non-git repos pre-submit, but tasks created via other entry points (CLI,
- * adopted state) can still reach here with a bare `fatal: not a git
- * repository`. Translate that into the real reason — a task is a git
- * worktree + branch, so for now the project must already be a git repo
- * (non-git roots are a planned follow-up) — instead of leaking git's stderr.
- */
-function worktreeErrorToast(err: unknown): string {
-  const raw = err instanceof Error ? err.message : String(err)
-  if (/not a git repository/i.test(raw)) {
-    return t("tasks.toast.worktreeErrorNotGit")
-  }
-  return t("tasks.toast.worktreeErrorGeneric", { message: raw })
-}
-
-function TasksShell(props: {
+export function TasksShell(props: {
   tasks: Accessor<readonly Task[]>
   initialTaskId?: string
   /**
@@ -154,9 +90,6 @@ function TasksShell(props: {
   // on. Falls back to `selectedId()` before the first cursor push.
   const [cursorId, setCursorId] = createSignal<string | null>(null)
   const actionTargetId = (): string | null => cursorId() ?? selectedId()
-  // Monotonic counter so a newer `switchTo` supersedes a slower in-flight one
-  // (see switchTo). Plain mutable int — nothing reactive observes it.
-  let switchToken = 0
 
   // Surface a user-action FAILURE as a red error toast. Under tmux's
   // alternate screen a bare `console.error` is invisible (it only reaches
@@ -196,12 +129,10 @@ function TasksShell(props: {
   }
   const [updateInfo, setUpdateInfo] = createSignal<UpdateInfo | null>(null)
 
-  // Zen-mode indicator: poll THIS pane's window for `@kobe_zen_panes` (set by
-  // the zen-toggle layout action while the ChatTab is collapsed to the engine
-  // pane). The Sidebar renders a `☯ ZEN` badge bottom-left when active. A
-  // 1s poll is cheap (one `show-options`, like the prefix probe) and there's
-  // no daemon channel for tmux-local layout state. No-op for a standalone
-  // `kobe tasks` with no `$TMUX_PANE`.
+  // Zen-mode indicator: poll THIS pane's window for `@kobe_zen_panes` (set
+  // by the zen-toggle layout action while the ChatTab is collapsed) so the
+  // Sidebar can show a `☯ ZEN` badge. No daemon channel for tmux-local
+  // layout state, so a cheap 1s poll; no-op with no `$TMUX_PANE`.
   const [zenActive, setZenActive] = createSignal(false)
   {
     const pane = process.env.TMUX_PANE
@@ -228,19 +159,13 @@ function TasksShell(props: {
   // pane on every resize.
   const dimensions = useTerminalDimensions()
 
-  // Shared active-task focus (the daemon's `active-task` channel, KOB-247) is
-  // followed ONLY by a HOME pane (no initialTaskId) — the navigational surface
-  // where selecting a row IS the focus, so it should mirror whatever was last
-  // entered from any session.
-  //
-  // A pane SPAWNED for a task session (initialTaskId set) represents exactly
-  // ONE task: its engine/chat is pinned to that task and never swaps. Following
-  // shared focus made such a pane LIE — click a sibling project to jump to it
-  // and this (now backgrounded) pane's highlight stuck on the sibling while its
-  // chat still showed its own task. So a task-bound pane pins its highlight to
-  // its own task and does NOT follow the channel; entering another task is a
-  // jump (switchTo), not a re-selection of THIS pane. (Origin: KOB-247 scoped
-  // shared focus to all panes; this narrows it back to the home pane.)
+  // Shared active-task focus (`active-task` channel, KOB-247) is followed
+  // ONLY by a HOME pane (no initialTaskId) — the navigational surface where
+  // selecting a row IS the focus. A pane SPAWNED for a task session pins its
+  // highlight to that one task instead: following shared focus made it LIE
+  // (a sibling click left this backgrounded pane's highlight on the sibling
+  // while its chat still showed its own task) — entering another task from
+  // there is a jump (switchTo), not a re-selection of THIS pane.
   createEffect(() => {
     if (props.initialTaskId) return
     const active = props.orch?.activeTaskSignal()()
@@ -248,16 +173,11 @@ function TasksShell(props: {
     setSelectedId(active)
   })
 
-  // Visual prefs (theme / transparent / focus accent) are applied
-  // centrally — boot + live `ui-prefs` pushes — by host-boot's
-  // UiPrefsSync; this shell no longer re-applies them itself. Sort mode,
-  // project filter, and the keys-legend fold ride the SAME `ui-prefs`
-  // channel but are pane-state, not theme-state, so the Tasks pane follows
-  // them here: a `t` / `ctrl+p` / `?` toggle in ANY session lands as a push
-  // and re-applies to this pane too. Changed-only assignment makes the echo
-  // of our own write (round-tripped through the watcher) a no-op. (Sort +
-  // project filter here; the keys-legend effect lives next to its signal
-  // further down.)
+  // Visual prefs apply centrally via host-boot's UiPrefsSync. Sort mode +
+  // project filter ride the SAME `ui-prefs` channel but are pane-state, so
+  // this pane follows them here: a toggle in ANY session pushes and
+  // re-applies here too. Changed-only assignment makes our own write's
+  // echo a no-op. (Keys-legend fold effect lives by its own signal below.)
   createEffect(
     on(
       () => props.orch?.uiPrefsSignal()(),
@@ -276,70 +196,36 @@ function TasksShell(props: {
     if (info) setUpdateInfo(info)
   })
 
-  // Shared task-action context (lib/task-actions). The flow bodies —
-  // confirm copy, DIRTY_WORKTREE force-delete branch, error handling —
-  // live in the shared module so this pane and the deprecated outer
-  // monitor can't drift apart. What stays here is only what's genuinely
-  // this host's: dialog wiring, toast surfacing, disk-only persistence,
-  // the chattab create surface, and selection.
-  const taskActions: CreateTaskContext = {
-    orch: props.orch,
+  // Deps bag for every action in `tasks-pane/actions.ts` (file-size cap) —
+  // built once per render, mirroring the settings-dialog/actions.ts
+  // template. Thin wrappers below keep every JSX prop / keybinding call
+  // site unchanged.
+  const actionsCtx: TasksHostActionsContext = {
     tasks: () => props.tasks(),
-    confirm: async (p) => (await DialogConfirm.show(dialog, p.title, p.body, p.cancelLabel, p.confirmLabel)) === true,
-    // Dialogs show IN the Tasks pane without zooming it full-window
-    // (KOB-244): the old `resize-pane -Z` hid the claude / ops / shell
-    // panes for the dialog's lifetime, which felt like the whole layout
-    // "popped out". The dialog overlay already caps to the pane width
-    // (`maxWidth = dimensions().width - 2`), so it renders fine in the
-    // ~22%-wide pane — just narrower — and the other panes stay visible.
-    promptText: (initial, opts) => RenameTaskDialog.show(dialog, initial, opts),
-    logger: console,
-    logPrefix: "[kobe tasks]",
+    orch: props.orch,
+    kv,
+    dialog,
     notifyError,
     notifyInfo,
     reload: () => props.reload(),
-    // This pane runs INSIDE the tmux client whose session a delete kills —
-    // switch away first so the kill doesn't yank the user's terminal.
-    switchBeforeKill: true,
-    // Publish the shared active-task focus so every surface follows (KOB-247).
-    updateActiveTask: true,
-    onTaskDeleted: (taskId, nextTask) => {
-      if (selectedId() !== taskId) return
-      const remaining = props.tasks()
-      setSelectedId(nextTask?.id ?? (remaining.find((t) => !t.archived) ?? remaining[0])?.id ?? null)
-    },
-    promptNewTask: (defaultRepo, repos, opts) => NewTaskDialog.show(dialog, defaultRepo, repos, opts),
-    // "Spawn a sibling" default: the cursor task's repo (fallback: the
-    // first listed task's).
-    cursorRepo: () => {
-      const list = props.tasks()
-      return (list.find((t) => t.id === selectedId()) ?? list[0])?.repo
-    },
-    // This pane uses disk-only persistence (no in-process kv store), so the
-    // atomic disk write is sufficient — no onRepoSaved kv mirror needed.
-    lastVendor: () => resolvePersistedVendor(getPersistedString("lastSelectedVendor"), getCustomEngineIds()),
-    rememberVendor: (vendor) => setPersistedString("lastSelectedVendor", vendor),
-    // Same surface preference as Settings (default chattab): open the
-    // new-task flow as a dedicated full-window page in a new tmux tab.
-    // The page performs the create/adopt itself and the subscribe pushes
-    // the new task back into this list. Fall back to the in-pane overlay
-    // if we can't resolve our tmux session.
-    openCreateSurface: async (defaultRepo) => {
-      const surface = normalizeSettingsSurface(kv.get(SETTINGS_SURFACE_KEY, DEFAULT_SETTINGS_SURFACE))
-      if (surface !== "chattab") return false
-      const session = await currentSessionName()
-      if (!session) return false
-      await openNewTaskTab(session, defaultRepo)
-      return true
-    },
-    // Land the cursor on the new task so Enter / click enters it next.
-    // (The daemon subscribe pushes the new task into the list momentarily.)
-    selectTask: (id) => setSelectedId(id),
-    // Then enter it: `n` drops the user straight into the new task's engine
-    // pane (the same enter loop Enter/click runs), ready to type the first
-    // prompt — not just a moved cursor.
-    enterTask: (id) => switchTo(id),
+    updateInfo,
+    setSelectedId,
   }
+  const switchRef: SwitchToRef = { token: 0 }
+  const switchTo = (id: string): Promise<void> => switchToAction(actionsCtx, switchRef, id)
+
+  // Shared task-action context (lib/task-actions): the flow bodies (confirm
+  // copy, DIRTY_WORKTREE force-delete branch, error handling) live in that
+  // shared module so this pane and the deprecated outer monitor can't drift
+  // apart; `buildTaskActionsContext` supplies only what's genuinely this
+  // host's (dialog wiring, toast surfacing, disk-only persistence, the
+  // chattab create surface, and selection).
+  const taskActions: CreateTaskContext = buildTaskActionsContext({
+    ...actionsCtx,
+    selectedId,
+    setSelectedId,
+    switchTo,
+  })
 
   // `n` creates a new task using the SAME NewTaskDialog (and now the same
   // createTaskFlow) as the outer app — parity matters; this pane is meant
@@ -348,57 +234,6 @@ function TasksShell(props: {
   // in-process. Backlog task (worktree lazy on first enter).
   async function createTask(): Promise<void> {
     await createTaskFlow(taskActions)
-  }
-
-  // Settings opens on the user's chosen surface (default chattab): a
-  // dedicated full-window `kobe settings` page opened as a new tmux tab,
-  // or the in-pane SettingsDialog overlay. If we can't resolve our tmux
-  // session (e.g. running outside a kobe pane), fall back to the overlay.
-  async function openSettings(): Promise<void> {
-    const surface = normalizeSettingsSurface(kv.get(SETTINGS_SURFACE_KEY, DEFAULT_SETTINGS_SURFACE))
-    if (surface === "chattab") {
-      const session = await currentSessionName()
-      if (session) {
-        await openSettingsTab(session)
-        return
-      }
-    }
-    const result = await SettingsDialog.show(dialog, kv, props.orch ?? undefined)
-    if (!result.visualPrefsChanged) return
-    if (!kv.flush()) return
-    try {
-      const session = await currentSessionName()
-      if (session) await refreshKobeWorkspacePanes(session)
-    } catch (err) {
-      console.error("[kobe tasks] failed to refresh workspace panes:", err)
-    }
-  }
-
-  // F1 help opens as a dedicated full-window tab (like Settings) — the
-  // in-pane HelpDialog overlay only had the narrow Tasks rail to render
-  // in, which truncated every keybinding row. Fall back to the overlay
-  // when we can't resolve our tmux session (e.g. running outside a kobe
-  // pane).
-  async function openHelp(): Promise<void> {
-    const session = await currentSessionName()
-    if (session) {
-      await openHelpTab(session)
-      return
-    }
-    HelpDialog.show(dialog)
-  }
-
-  async function openUpdate(): Promise<void> {
-    const info = updateInfo()
-    if (!info?.hasUpdate) {
-      // The `u` chord / update chip would otherwise no-op silently when
-      // nothing is pending — confirm the up-to-date state instead (#23a).
-      notifyInfo(t("tasks.toast.alreadyLatest", { version: CURRENT_VERSION }))
-      return
-    }
-    const session = await currentSessionName()
-    if (!session) return
-    await openUpdateTab(session)
   }
 
   async function archiveTask(id: string): Promise<void> {
@@ -426,98 +261,15 @@ function TasksShell(props: {
     await cycleVendorFlow(taskActions, id)
   }
 
-  async function openSelectedWorktree(id: string): Promise<void> {
-    const task = props.tasks().find((t) => t.id === id)
-    let worktree = task?.worktreePath
-    if (!worktree || !existsSync(worktree)) {
-      if (!props.orch) {
-        console.error("[kobe tasks] no daemon; cannot materialise worktree")
-        notifyError(t("tasks.toast.noDaemonWorktree"))
-        return
-      }
-      try {
-        worktree = await props.orch.ensureWorktree(id)
-      } catch (err) {
-        console.error("[kobe tasks] task.ensureWorktree failed:", err)
-        notifyError(worktreeErrorToast(err))
-        return
-      }
-      await props.reload()
-    }
-    if (!worktree || !existsSync(worktree)) return
-    const opener = detectWorktreeOpener()
-    if (!opener) {
-      console.error("[kobe tasks] no editor/opener found; set KOBE_OPEN_EDITOR")
-      notifyError(t("tasks.toast.noEditor"))
-      return
-    }
-    if (!openWorktree(worktree, opener)) {
-      console.error(`[kobe tasks] failed to open worktree with ${opener.label}`)
-      notifyError(t("tasks.toast.openWorktreeFailed", { label: opener.label }))
-    }
-  }
-
-  // Right arrow → re-focus THIS window's engine (claude/codex) pane, the
-  // inverse of ctrl+h. Targeting: the role-tagged lookup
-  // (`claudePaneIdStrict` → `@kobe_role=claude`, vendor-neutral) is the
-  // honest "current window's engine pane" — `paneIdByRole` lists the
-  // session's ACTIVE window, which is necessarily the window holding this
-  // pane (we only receive the keystroke while active), so no window
-  // derivation from $TMUX_PANE is needed. Preferred over
-  // `select-pane -t $TMUX_PANE -R`, which grabs whatever pane happens to
-  // sit right of the rail (wrong after a manual rearrange) — same shape as
-  // selectTasksPane, the ctrl+q first stage, just pointed the other way.
-  // No-op outside tmux (standalone `kobe tasks` has no $TMUX_PANE) or when
-  // the window has no tagged engine pane (legacy session).
-  async function focusEnginePane(): Promise<void> {
-    if (!process.env.TMUX_PANE) return
-    const session = await currentSessionName()
-    if (!session) return
-    const pane = await claudePaneIdStrict(session)
-    if (!pane) return
-    await runTmux(["select-pane", "-t", pane])
-  }
-
-  async function moveTask(id: string, delta: -1 | 1): Promise<void> {
-    const task = props.tasks().find((t) => t.id === id)
-    if (!task || task.kind === "main" || !props.orch) return
-    try {
-      await props.orch.moveTask(id, delta)
-    } catch (err) {
-      console.error("[kobe tasks] task.move failed:", err)
-      notifyError(t("tasks.toast.moveTaskFailed", { message: err instanceof Error ? err.message : String(err) }))
-      return
-    }
-    setSelectedId(id)
-    await props.reload()
-  }
-
-  // Pin / unpin the cursor task (Shift+P). `setPinned` with no explicit flag
-  // toggles daemon-side; reload mirrors the new order into the poll fallback,
-  // matching how archive/delete refresh the list. No-op without a daemon.
-  // A pin failure is log-only (like focusEnginePane): it's rare, non-destructive,
-  // and there's no dedicated toast key in the tasks-pane message namespace.
-  async function togglePin(id: string): Promise<void> {
-    if (!props.orch) return
-    try {
-      await props.orch.setPinned(id)
-    } catch (err) {
-      console.error("[kobe tasks] task.pin failed:", err)
-      return
-    }
-    await props.reload()
-  }
-
-  // `i` — flip a task between the live read-only preview and the engine. The
-  // mode is read when the session is built and a healthy session is reused
-  // as-is, so kill this task's session to force a rebuild in the new mode, then
-  // switch into it to show the result. killSession no-ops when none exists.
-  async function togglePreviewFlow(id: string): Promise<void> {
-    const { togglePreviewMode } = await import("@/state/preview-mode")
-    togglePreviewMode(id)
-    await killSession(tmuxSessionName(id))
-    await switchTo(id)
-  }
+  const openSettings = (): Promise<void> => openSettingsAction(actionsCtx)
+  const openHelp = (): Promise<void> => openHelpAction(actionsCtx)
+  const openWorktrees = (): Promise<void> => openWorktreesAction()
+  const openUpdate = (): Promise<void> => openUpdateAction(actionsCtx)
+  const openSelectedWorktree = (id: string): Promise<void> => openSelectedWorktreeAction(actionsCtx, id)
+  const focusEnginePane = (): Promise<void> => focusEnginePaneAction()
+  const moveTask = (id: string, delta: -1 | 1): Promise<void> => moveTaskAction(actionsCtx, id, delta)
+  const togglePin = (id: string): Promise<void> => togglePinAction(actionsCtx, id)
+  const togglePreviewFlow = (id: string): Promise<void> => togglePreviewFlowAction(actionsCtx, switchRef, id)
 
   // Collapsed state of the `── keys ──` legend (toggled by `?` / clicking
   // the header). A GLOBAL pref fanned out like sort/theme: seed from the
@@ -554,15 +306,16 @@ function TasksShell(props: {
   useBindings(() => ({
     enabled: dialog.stack.length === 0 && !searchActive(),
     // Chords come from KobeKeymap via bindByIds (f1=help.open, n=task.new,
-    // s=settings.open.sidebar, u=tasks.update, o/b/v=tasks.*) so user
-    // overrides from ~/.kobe/settings/keybindings.yaml apply here too.
-    // F1 → the full-window help tab (openHelp), with the shared HelpDialog
-    // as the no-session fallback. Gated on an empty dialog stack so it
-    // doesn't `replace` an open dialog.
+    // s=settings.open.sidebar, w=worktrees.open.sidebar, u=tasks.update,
+    // o/b/v=tasks.*) so user overrides from ~/.kobe/settings/keybindings.yaml
+    // apply here too. F1 → the full-window help tab (openHelp), with the
+    // shared HelpDialog as the no-session fallback. Gated on an empty
+    // dialog stack so it doesn't `replace` an open dialog.
     bindings: bindByIds({
       "help.open": () => void openHelp(),
       "task.new": () => void createTask(),
       "settings.open.sidebar": () => void openSettings(),
+      "worktrees.open.sidebar": () => void openWorktrees(),
       "tasks.update": () => void openUpdate(),
       "tasks.openWorktree": () => {
         const id = actionTargetId()
@@ -583,51 +336,6 @@ function TasksShell(props: {
         void focusEnginePane().catch((err) => console.error("[kobe tasks] focus engine pane failed:", err)),
     }),
   }))
-
-  // Enter / click on a task → switch this tmux client to that task's
-  // session, creating it on demand. The full enter loop:
-  //   1. session running → just switch-client.
-  //   2. session gone but worktree on disk → ensureSession + switch.
-  //   3. backlog task, no worktree yet → materialise it via the daemon's
-  //      `task.ensureWorktree` RPC (git worktree add — only the
-  //      Orchestrator can do it), then ensureSession + switch.
-  // Step 3 closes the create→enter loop entirely inside the Tasks pane
-  // (a task you just made with `n` is enterable here, no detour through
-  // the outer monitor).
-  async function switchTo(id: string): Promise<void> {
-    const task = props.tasks().find((t) => t.id === id)
-    if (!task) return
-    // Monotonic switch token: a newer switchTo supersedes any still-in-flight
-    // one. enterTask checks `isCurrent` before its disruptive setActive +
-    // switch-client, so a slow cold-session switch that finishes after a later
-    // click can't drag the active task (and thus the home pane's selection)
-    // back to the superseded project. (Fixes the "click A → click B → top-left
-    // selection sticks on A" race.)
-    const myToken = ++switchToken
-    // The whole enter sequence (capture from-layout → ensure/heal session →
-    // zen → setActive → fit + switch) lives in the Handover owner; the Tasks
-    // pane opts into capture + heal and reloads its mirror after a cold
-    // worktree materialise. `includeInitPrompt: true` keeps prior switchTo
-    // behaviour (always weave the marker-guarded repo-init). enterTask throws a
-    // HandoverError on a build failure; we map its phase to the right toast.
-    try {
-      await enterTask(props.orch, task, task.repo, task.vendor, {
-        includeInitPrompt: true,
-        heal: true,
-        captureFrom: true,
-        reload: () => props.reload(),
-        isCurrent: () => switchToken === myToken,
-      })
-    } catch (err) {
-      if (err instanceof HandoverError) {
-        if (err.phase === "no-daemon") notifyError(t("tasks.toast.noDaemonOpen"))
-        else if (err.phase === "worktree") notifyError(worktreeErrorToast(err.cause ?? err))
-        else notifyError(t("tasks.toast.sessionStartFailed"))
-      } else {
-        console.error("[kobe tasks] switchTo failed:", err)
-      }
-    }
-  }
 
   // Version / update chip for the Sidebar's brand header — moved up from the
   // footer's old `── system ──` block. Emphasised + clickable (opens the
@@ -774,256 +482,6 @@ function TasksShell(props: {
   )
 }
 
-/**
- * Resolve a single binding id to the chord cap the footer should advertise:
- * the cosmetic `hint.keys` when present (it's refreshed in place on an
- * override — keymap-overrides.ts), else the canonical first chord. Returns
- * `null` when the id is unbound (no chords) — the row that owns it should
- * then drop, since advertising a dead chord is worse than none (mirrors the
- * override path that nulls a hint on unbind).
- *
- * Pure + exported so the legend derivation is unit-testable against a faked
- * keymap without booting a tmux pane (the host itself isn't CI-runnable).
- */
-export function legendCap(id: string): string | null {
-  const row = findBinding(id)
-  if (!row) return null
-  const cap = row.hint?.keys ?? row.keys[0]
-  return cap && cap.length > 0 ? cap : null
-}
-
-/**
- * Resolve a (possibly composite) legend row's keycap from the binding ids it
- * represents. Each id contributes its {@link legendCap}; unbound ids drop out
- * and the survivors join with `/` (so `r/b/v` becomes `r/v` if `b` is
- * unbound, or the whole row drops when nothing survives). Returns `null` when
- * every id resolved to no chord — the caller drops the row entirely.
- */
-export function legendRowCap(ids: readonly string[]): string | null {
-  const caps = ids.map(legendCap).filter((c): c is string => c !== null)
-  return caps.length > 0 ? caps.join("/") : null
-}
-
-/**
- * A small shortcut legend pinned to the bottom of the Tasks pane (KOB-244):
- * shows the in-pane task actions plus the session-level tmux chords so the
- * keys are discoverable without leaving the pane. The `ctrl+h/j/k/l` and
- * `ctrl+[/]` lines are tmux session bindings — shown here, not rebound.
- *
- * Collapsible: the legend is ~20 rows with the tmux chords included, which
- * crowds the task list on short terminals. `?` (or clicking the header)
- * folds it down to the header line; move-mode hints ignore the fold — a
- * user inside reorder mode must always see how to leave it.
- */
-function ShortcutHints(props: {
-  moveMode?: Accessor<boolean>
-  selectedIsMain?: Accessor<boolean>
-  collapsed?: Accessor<boolean>
-  onToggleCollapsed?: () => void
-}) {
-  const { theme } = useTheme()
-  // Resolve the user's REAL tmux prefix at runtime (#12). kobe loads the
-  // user's own prefix, so a literal `Prefix F` is un-actionable — the user may
-  // not know their prefix is C-a. Shell `tmux show-options -g prefix` on the
-  // -L kobe socket (runTmuxCapturing already targets it) and render `C-b` as
-  // `⌃B`. Falls back to the literal `Prefix` when resolution fails / is flaky.
-  const [prefixCap, setPrefixCap] = createSignal("Prefix")
-  onMount(() => {
-    void runTmuxCapturing(["show-options", "-g", "prefix"]).then(({ code, stdout }) => {
-      if (code !== 0) return
-      const glyph = tmuxPrefixGlyph(stdout)
-      if (glyph) setPrefixCap(glyph)
-    })
-  })
-  // A hint row. `k` is a MACHINE chord string (`ctrl+q`, `prefix f`, `a/d`);
-  // it's rendered as macOS glyphs via `formatChord` at draw time, so the
-  // footer, the F1 help, and the status bar all read the same (one formatter,
-  // no drift). `dimWhenMain` flags a keycap whose action early-returns on a
-  // `main` (project root) row — the footer dims that cap so a press there reads
-  // as "doesn't apply here" rather than a silent no-op (Issue #7).
-  type Hint = { k: string; label: string; dimWhenMain?: boolean }
-  // tmux session-key rows derive from the RESOLVED key set so user
-  // overrides (`tmux.*` ids in ~/.kobe/settings/keybindings.yaml) show
-  // their own chords here; an unbound id drops its row. Pseudo-chords
-  // ("ctrl+hjkl", "ctrl+[/]") are kept only while the relevant keys are
-  // still at their defaults — overridden keys render as plain chords.
-  const tmuxHints = (): ReadonlyArray<Hint> => {
-    // Re-derive after a live keybindings reload: the bump invalidates this
-    // accessor so the footer re-renders with the freshly-resolved tmux
-    // chords (the resolver's cache is cleared in the same reload).
-    keymapVersion()
-    const res = resolveUserTmuxKeys()
-    const b = res.binds
-    const out: Hint[] = []
-    const focusChords = res.focus.filter((f): f is NonNullable<typeof f> => f !== null).map((f) => f.chord)
-    if (focusChords.length === 4 && focusChords.every((c, i) => c === TMUX_FOCUS_DEFAULTS[i])) {
-      out.push({ k: "ctrl+hjkl", label: t("tasks.hints.movePanes") })
-    } else if (focusChords.length > 0) {
-      out.push({ k: focusChords[0] as string, label: t("tasks.hints.movePanes") })
-    }
-    const layoutGroup = (label: string, ids: readonly (keyof typeof b)[]): void => {
-      const chords = ids.map((id) => b[id]?.chord).filter((chord): chord is string => !!chord)
-      if (chords.length > 0) out.push({ k: `prefix ${chords.join("/")}`, label })
-    }
-    // Trimmed legend: keep pane movement, the tasks→detach chord, and the two
-    // tmux-prefix layout groups. Per-tab rows (switch / new / engine / rename /
-    // close) live in F1 full help, not the footer.
-    if (b["tmux.detach"]) out.push({ k: b["tmux.detach"].chord, label: t("tasks.hints.detach") })
-    layoutGroup(t("tasks.hints.splits"), [
-      "tmux.layout.workspaceSplit",
-      "tmux.layout.workspaceClose",
-      "tmux.layout.workspaceReset",
-    ])
-    layoutGroup(t("tasks.hints.panes"), [
-      "tmux.layout.tasksToggle",
-      "tmux.layout.opsToggle",
-      "tmux.layout.terminalToggle",
-    ])
-    return out
-  }
-  // Fixed-width key column so labels line up — a terminal-grammar legend
-  // column, not a proportional pane (allowed hardcode). formatChord keeps
-  // plain-letter caps lowercase (the EXACT key to press, #14), uppercases the
-  // key of modifier chords (`⌃ Q`), shows `tab` as a word, and renders the two
-  // `prefix …` rows with the user's REAL resolved prefix (`prefixCap()`, #12).
-  // Derived (not a static const) so those rows re-render once the async prefix
-  // resolution lands.
-  // Each in-pane row's keycap is DERIVED from KobeKeymap (legendRowCap) so a
-  // user override / unbind in ~/.kobe/settings/keybindings.yaml is reflected
-  // here — the footer is the only always-visible legend, and the doc promises
-  // it follows the keymap (docs/KEYBINDINGS.md). The ids mirror a curated
-  // subset of the pane's bindings plus the Sidebar-owned sidebar.* rows it
-  // delegates to (Enter→sidebar.select, [/]→sidebar.view, d→sidebar.delete).
-  // `keymapVersion()` is read at the top so a live reload re-renders the
-  // legend with the freshly-resolved chords — same pattern as tmuxHints().
-  // Each row is conditional: an id that resolved to no chord (unbound) drops
-  // its row rather than advertising a dead key.
-  const defaultHints = (): ReadonlyArray<Hint> => {
-    keymapVersion()
-    // Trimmed legend (KOB request): the footer carries only the high-traffic
-    // rows; everything else (sort, move/merge, archive, rename/branch/engine,
-    // per-tab tmux chords) is reachable via F1 full help. Order here is the
-    // exact order the rows render in.
-    const rows: Array<{ ids: readonly string[]; label: string; dimWhenMain?: boolean }> = [
-      { ids: ["help.open"], label: t("tasks.hints.fullHelp") },
-      { ids: ["task.new"], label: t("tasks.hints.newTask") },
-      { ids: ["settings.open.sidebar"], label: t("tasks.hints.settings") },
-      { ids: ["sidebar.select"], label: t("tasks.hints.open") },
-      // Right arrow re-focuses the current window's engine pane
-      // (tasks.focusEngine) — renders as [→] via formatChord's KEY_GLYPH.
-      { ids: ["tasks.focusEngine"], label: t("tasks.hints.focusEngine") },
-      { ids: ["tasks.openWorktree"], label: t("tasks.hints.openWorktree") },
-      { ids: ["sidebar.delete"], label: t("tasks.hints.delete") },
-      { ids: ["sidebar.view"], label: t("tasks.hints.views") },
-      { ids: ["sidebar.projectFilter"], label: t("tasks.hints.project") },
-    ]
-    const out: Hint[] = []
-    for (const row of rows) {
-      const k = legendRowCap(row.ids)
-      if (k === null) continue
-      out.push({ k, label: row.label, dimWhenMain: row.dimWhenMain })
-    }
-    out.push(...tmuxHints())
-    return out
-  }
-  const MOVE_HINTS = (): ReadonlyArray<Hint> => [
-    { k: "j/k", label: t("tasks.hints.reorder") },
-    { k: "enter/esc", label: t("tasks.hints.done") },
-  ]
-  const hints = () => (props.moveMode?.() ? MOVE_HINTS() : defaultHints())
-  // Width of the description column = the longest label, but CAPPED so a long
-  // label can't blow the column out past what the 32-cell Tasks pane (minus the
-  // 10-cell keycap column) can hold. Each row right-aligns this fixed-width box
-  // (text left-aligned inside), so every description shares one left edge AND
-  // the whole column hugs the pane's right side. Labels longer than the cap are
-  // ellipsised rather than allowed to overflow.
-  const LABEL_COL_MAX = 18
-  // Size the column in DISPLAY CELLS, not code points: a CJK label (default zh)
-  // renders 2 cells per glyph, so `.length` (code points) sized the box at ~half
-  // the needed width and clipped the label. `approxCellWidth` (the CJK-aware
-  // helper the hover tooltip already uses) counts fullwidth codepoints as 2.
-  const labelColWidth = () => Math.min(LABEL_COL_MAX, Math.max(...hints().map((h) => approxCellWidth(h.label))))
-  // Truncate to the CELL budget too. truncateEnd's budget is code points, so a
-  // CJK label needs a code-point budget of at most floor(cells/2) to fit the
-  // cell-sized box; take the tighter of the two so a wide label can't overflow.
-  const clipLabel = (s: string): string => {
-    const cells = labelColWidth()
-    if (approxCellWidth(s) <= cells) return s
-    const points = [...s]
-    let used = 0
-    let cut = 0
-    for (const ch of points) {
-      const w = (ch.codePointAt(0) ?? 0) >= 0x1100 ? 2 : 1
-      if (used + w > cells - 1) break // reserve 1 cell for the ellipsis
-      used += w
-      cut++
-    }
-    return `${points.slice(0, cut).join("")}…`
-  }
-  // Version + update moved UP to the Sidebar's `kobe` brand header (the old
-  // `── system ──` block lived here); the footer is now just the key legend.
-  // Move-mode overrides the fold: its two hints are the only exit
-  // instructions for reorder mode, so they always render.
-  const folded = () => (props.collapsed?.() ?? false) && !(props.moveMode?.() ?? false)
-  return (
-    <box
-      flexShrink={0}
-      flexDirection="column"
-      paddingLeft={1}
-      paddingRight={1}
-      paddingTop={1}
-      paddingBottom={1}
-      gap={0}
-    >
-      {/* Header doubles as the toggle: `?` chord or a click folds/unfolds.
-          The `?▸ / ?▾` tail advertises both the chord and the state. */}
-      <text
-        fg={theme.textMuted}
-        attributes={TextAttributes.DIM}
-        wrapMode="none"
-        onMouseUp={() => props.onToggleCollapsed?.()}
-      >
-        {folded() ? t("tasks.hints.headerFolded") : t("tasks.hints.headerUnfolded")}
-      </text>
-      <Show when={!folded()}>
-        <For each={hints()}>
-          {(h) => {
-            // Dim a cap whose action early-returns on a `main` row (`B`/`M`):
-            // muted + DIM instead of bold accent, so the user sees it doesn't
-            // apply to the project row rather than pressing it into silence.
-            const dim = () => h.dimWhenMain === true && (props.selectedIsMain?.() ?? false)
-            return (
-              <box flexDirection="row" gap={1} justifyContent="space-between">
-                {/* `[key]` keycap chip — agent-deck style, mirrors the outer
-                monitor's StatusBar Hotkey: bold accent key in brackets,
-                muted label. No fill, so it stays clean in transparent mode. */}
-                <box width={10} flexShrink={0}>
-                  <text
-                    fg={dim() ? theme.textMuted : theme.accent}
-                    attributes={dim() ? TextAttributes.DIM : TextAttributes.BOLD}
-                    wrapMode="none"
-                  >
-                    [{formatChord(h.k, prefixCap())}]
-                  </text>
-                </box>
-                {/* Description column — fixed width = longest label, pushed to the
-                right edge by space-between. Text is left-aligned inside, so
-                every description shares one left edge while the whole column
-                hugs the right side and rides the pane width. */}
-                <box width={labelColWidth()} flexShrink={0}>
-                  <text fg={theme.textMuted} attributes={dim() ? TextAttributes.DIM : undefined} wrapMode="none">
-                    {clipLabel(h.label)}
-                  </text>
-                </box>
-              </box>
-            )
-          }}
-        </For>
-      </Show>
-    </box>
-  )
-}
-
 export async function startTasksPane(opts: { initialTaskId?: string } = {}): Promise<void> {
   await bootPaneHost({
     logContext: "tasks",
@@ -1036,112 +494,4 @@ export async function startTasksPane(opts: { initialTaskId?: string } = {}): Pro
     providers: { notifications: true },
     setup: () => setupTasksPane(opts),
   })
-}
-
-async function setupTasksPane(opts: { initialTaskId?: string }): Promise<HostScreen> {
-  // Task source. PRIMARY = a live daemon SUBSCRIBE (via RemoteOrchestrator):
-  // a task created / renamed / deleted in ANY session's Tasks pane or in
-  // the outer monitor is pushed to THIS pane in real time, so every
-  // session's list stays in sync (KOB-244 — a new task wasn't showing up
-  // in an already-open session's Tasks pane). The shared env baked onto
-  // this pane's command (inheritedEnvPrefix) guarantees we connect to the
-  // SAME daemon as everyone else.
-  //
-  // FALLBACK = a direct tasks.json read + slow poll, used only when the
-  // daemon is unreachable. MUST pass `homeDir()` (KOBE_HOME_DIR-aware) or
-  // it would read the PRODUCTION `~/.kobe/tasks.json` (KOB-233).
-  const store = new TaskIndexStore({ homeDir: homeDir() })
-  await store.load()
-  const [fileTasks, setFileTasks] = createSignal<readonly Task[]>(store.list())
-
-  let orch: RemoteOrchestrator | null = null
-  try {
-    // NON-spawning connect. A Tasks pane subscribes as role:"pane" and must
-    // NEVER start a daemon — doing so would resurrect an idle-stopped daemon
-    // with no gui to hold it, breaking the refcounted lazy-shutdown. This bit
-    // most visibly via `kobe reload`, which respawns this pane while the user
-    // may be detached (daemon already idle-stopped): a spawning connect would
-    // leave a gui-less daemon running forever. A gui owns daemon lifecycle; if
-    // none is up we fall through to the always-on tasks.json poll below.
-    const client = await connectIfRunning()
-    if (client) {
-      const remote = new RemoteOrchestrator(client)
-      await remote.init() // hello + subscribe → tasksSignal() is now live
-      orch = remote
-    } else {
-      logClient("tasks-boot", "no daemon running — polling tasks.json (a gui owns daemon lifecycle)")
-    }
-  } catch (err) {
-    logClientError("tasks-boot", err)
-    logClient("tasks-boot", "daemon subscribe failed — polling tasks.json")
-  }
-
-  // Display source: prefer the daemon's live snapshot WHILE the socket is
-  // online, otherwise fall back to the file poll. A plain-function accessor
-  // (not createMemo) so it isn't a computation created outside a render root;
-  // it reactively tracks whichever signals it reads on each call. The crucial
-  // fix for the create/delete sync drift: when the daemon idle-stops / restarts
-  // and the socket closes, `connectionStateSignal()` flips to "disconnected"
-  // and the display switches to the always-running file poll instead of
-  // FREEZING on the last daemon snapshot (the old `orch ? orch.tasksSignal()`
-  // had no fallback once subscribed). The orchestrator's own non-spawning
-  // reconnect loop then restores the live path when a daemon returns.
-  const tasks: Accessor<readonly Task[]> = () =>
-    orch && orch.connectionStateSignal()() === "online" ? orch.tasksSignal()() : fileTasks()
-  const reload = async (): Promise<void> => {
-    await store.load()
-    setFileTasks(store.list())
-  }
-  // ALWAYS run the backstop poll (not gated on daemon availability, unlike
-  // before — that gate was the freeze bug). It does the file read only when
-  // the daemon push path is NOT the live source, so an online pane pays
-  // nothing and an offline one stays fresh within RELOAD_MS.
-  //
-  // Offline ticks are additionally mtime-gated (waste audit): tasks.json
-  // only changes on a mutation, so a cheap `stat` decides whether the full
-  // read+parse is needed — an idle offline pane pays one stat per 1.5s
-  // instead of re-reading and re-parsing the whole index 40×/min. Writes
-  // are atomic temp+rename, so mtime+size always move on a real change.
-  // A stat failure maps to a distinct "missing" fingerprint: deletion →
-  // recreation each reload exactly once. Explicit `reload()` calls (after
-  // mutations) bypass the gate on purpose. Errors are swallowed — this
-  // pane process has no crash net (see ops/host.tsx), so a transient fs
-  // error must degrade to a stale list, not an unhandled rejection.
-  let lastTasksFileFingerprint = ""
-  const timer = setInterval(() => {
-    if (orch && orch.connectionStateSignal()() === "online") return
-    void (async () => {
-      // Detached (background) session: skip even the stat — nobody's looking.
-      if (!(await sessionAttached())) return
-      let fingerprint = "missing"
-      try {
-        const st = await stat(store.filePath)
-        fingerprint = `${st.mtimeMs}:${st.size}`
-      } catch {
-        // keep the "missing" fingerprint
-      }
-      if (fingerprint === lastTasksFileFingerprint) return
-      lastTasksFileFingerprint = fingerprint
-      await reload()
-    })().catch(() => {})
-  }, RELOAD_MS)
-
-  return {
-    root: () => (
-      <>
-        <TasksShell tasks={tasks} initialTaskId={opts.initialTaskId} orch={orch} reload={reload} />
-        <ToastOverlay />
-      </>
-    ),
-    // Tear down on ACTUAL exit, not after render() resolves: `render`
-    // resolves at mount (cf. startApp, which also cleans up via
-    // onDestroy), so disposing here is the only correct place. Disposing
-    // after `await render(...)` killed the daemon client + poll the moment
-    // the pane mounted → "daemon client disposed" on the next switch and
-    // a dead subscribe (KOB-247).
-    onDestroy: () => {
-      if (timer) clearInterval(timer)
-      orch?.dispose()
-    },
-  }
 }
